feat: Add knowledge base support with vector indexing and gRPC APIs
- Add knowledge base table schema (KnowledgeBaseInfo, KnowledgeBaseEntry)
- Implement one-click vector index generation for knowledge bases
- Add comprehensive gRPC APIs for knowledge base management
- Add knowledge base CRUD operations in yakit package
- Add vector indexing functions for both full knowledge base and individual entries
- Add CI tests for knowledge base functionality with MUSTPASS prefix
- Support search, pagination, and filtering for knowledge base entries
- Integrate with RAG system for semantic search capabilities

+++ common/ai/rag/rag_test.go
@@ -165,6 +165,6 @@ func TestFilterResults(t *testing.T) {
 
 	filtered := FilterResults(results, 0.6)
 	assert.Equal(t, 2, len(filtered))
-	assert.Equal(t, 0.9, filtered[0].Score)
-	assert.Equal(t, 0.7, filtered[1].Score)
+	assert.Equal(t, float32(0.9), filtered[0].Score)
+	assert.Equal(t, float32(0.7), filtered[1].Score)
 }
